+++ src/tribler-core/tribler_core/requirements.txt
@@ -21,3 +21,5 @@ sentry-sdk==1.5.0
 service-identity==21.1.0
 yappi==1.3.3
 yarl==1.7.2 # keep this dependency higher than 1.6.3. See: https://github.com/aio-libs/yarl/issues/517
+Faker==9.8.2
+sentry-sdk==1.5.0
