+++ src/tribler-gui/tribler_gui/tribler_window.py
@@ -41,7 +41,7 @@ from psutil import LINUX
 from tribler_common.network_utils import default_network_utils
 from tribler_common.process_checker import ProcessChecker
 from tribler_common.rest_utils import FILE_SCHEME, MAGNET_SCHEME, scheme_from_uri, uri_is_valid_file, uri_to_path
-from tribler_common.utilities import parse_query
+from tribler_core.utilities.utilities import parse_query
 from tribler_common.version_manager import VersionHistory
 
 from tribler_core.utilities.unicode import hexlify
