+++ src/tribler-gui/tribler_gui/widgets/tablecontentmodel.py
@@ -8,7 +8,7 @@ from typing import Callable, Dict
 from PyQt5.QtCore import QAbstractTableModel, QModelIndex, QRectF, QSize, Qt, pyqtSignal
 
 from tribler_common.simpledefs import CHANNELS_VIEW_UUID, CHANNEL_STATE
-from tribler_common.utilities import to_fts_query
+from tribler_core.utilities.utilities import to_fts_query
 
 from tribler_core.components.metadata_store.db.orm_bindings.channel_node import NEW
 from tribler_core.components.metadata_store.db.serialization import CHANNEL_TORRENT, COLLECTION_NODE, REGULAR_TORRENT
